refactor: centralize token pruning helper
